requirements.txt
flask
redis
pyjwt
requests
meinheld